Enhancement by mentor

+++ include/processor.h
@@ -3,7 +3,7 @@
 
 class Processor {
  public:
-  float Utilization();  // TODO: See src/processor.cpp
+  double Utilization();  // TODO: See src/processor.cpp
 
   // TODO: Declare any necessary private members
  private:
